Move fedifyMiddleware() callbacks to handlers.ts
Move the onRequest and onBeforeResponse implementations that lived
inline in fedifyMiddleware() into src/handlers.ts, keeping their
behavior unchanged.

This prepares for testing the handlers independently of SolidStart.
Taking fedifyMiddleware() itself as the SUT would require groundwork
for loading the SolidStart framework into the test runtime, and no
clean way to do that turned up.
I judged that extracting the two callbacks that made up
fedifyMiddleware() and taking them as the SUT instead would still
cover the behavior.
